requirements.txt
langchain
python-dotenv
streamlit
torch
